pyproject.toml
[project]
name = "aoc25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "types-requests>=2.32.4.20250913",
    "types-tqdm>=4.67.0.20250809",
]
